Create a test template for a distribution

When you add a new distribution, you want to add a test to make sure it is behaving as expected. This function generates a test template. It only creates the text, it does not write to file. See write_distribution_test() to write the test to file automatically.

Usage

greta_distribution_template_test(dist_name)

Arguments

dist_name

character. A distribution name. E.g., "lognormal"

Value

text containing test template code

Examples

gumbel <- greta_distribution_template_test("gumbel")